Fixed OpenCL & CUDA using incorrect periodic box vectors
OpenMM 8.2.0 added a check to see if the periodic box is changed from
rectangular to triclinic after the OpenCL or CUDA platforms are created.
This was to prevent a bug where the platforms setup was performed with a
different box shape to what is used.

https://github.com/openmm/openmm/issues/4548

Now the code for setting the system periodic box vectors has been moved
so it takes place before the platform is created.